+++ apps/partial-prerendering/src/app/sandbox/client-static.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import React from "react";
+
+export function ClientStatic(props: React.PropsWithChildren<{}>) {
+  return (
+    <span>
+      {props.children ?? "ClientStatic"}
+    </span>
+  );
+}