+++ apps/partial-prerendering/src/app/sandbox/server-static.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export function ServerStatic(props: React.PropsWithChildren<{}>) {
+  return (
+    <span>
+      {props.children ?? "ServerStatic"}
+    </span>
+  );
+}