+++ apps/partial-prerendering/src/app/sandbox/simple-suspense.tsx
@@ -1,0 +1,11 @@
+import { Suspense } from "react";
+
+export function SimpleSuspense(props: React.PropsWithChildren<{}>) {
+  return (
+    <>
+      Suspense{"<"}<Suspense fallback={<span>Loading...</span>}>
+        {props.children}
+      </Suspense>{">"}
+    </>
+  );
+}