+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,21 @@
 # demo-app
 
+## 1.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.3.4
+  - @stackframe/stack@2.3.6
+
+## 1.0.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.3.3
+  - @stackframe/stack@2.3.5
+
 ## 1.0.20
 
 ### Patch Changes