+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # stack-docs
 
+## 0.0.4
+
+### Patch Changes
+
+- Partial pre-rendering
+
 ## 0.0.3
 
 ### Patch Changes