+++ packages/stack-server/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-server
 
+## 0.3.6
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.3.4
+  - @stackframe/stack@2.3.6
+
+## 0.3.5
+
+### Patch Changes
+
+- Partial pre-rendering
+- Updated dependencies
+  - @stackframe/stack-shared@2.3.3
+  - @stackframe/stack@2.3.5
+
 ## 0.3.4
 
 ### Patch Changes